Renamed categories "serializer" to "serializers"
This fixes #318.
